# Conflicts:
#	Makefile.x86
#	patches/startup/startup.sh
#	rp-api/api-hw-calib/src/calib.c

+++ apps-tools/calib_app/js/manager_adc_mode.js
@@ -512,7 +512,7 @@ $(function() {
         }
 
         if (_mode == "CH3_ADC_GAIN") {
-            CLIENT.parametersCache["ch3_gain_adc_new"] = { value: parseInt($("#CH3_GAIN").val()) + _new_val };
+            CLIENT.parametersCache["ch3_gain_adc_new"] = { value: parseFloat($("#CH3_GAIN").val()) + _new_val };
             CLIENT.sendParameters();
             OBJ.adcCalibChange = true;
         }
